backend/requirements.txt
websockets>=12.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-ApplicationServices>=10.0
google-generativeai>=0.8.0
openai>=1.0.0
openwakeword>=0.5.0
pyaudio>=0.2.14
numpy>=1.24.0
pyautogui>=0.9.54
Pillow>=10.0.0
flask>=3.0.0
